+++ build.gradle
@@ -1,14 +1,19 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    ext.kotlin_version = '1.9.24'
+    ext {
+        kotlin_version = '2.0.0'
+        about_libs = '11.2.2'
+    }
     repositories {
         google()
         mavenCentral()
+        maven { url "https://plugins.gradle.org/m2/" }
     }
     dependencies {
         classpath 'com.android.tools.build:gradle:8.2.0'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
+        classpath "com.mikepenz.aboutlibraries.plugin:aboutlibraries-plugin:$about_libs"
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
